requirements.txt
# Backend Requirements - Modern Stack
# Web Framework
flask==2.3.3
flask-cors==4.0.0

# Data Processing & ML
pandas==2.0.3
joblib==1.3.2
scikit-learn==1.3.0
numpy==1.24.3

# Database
oracledb==2.0.1
sqlalchemy==2.0.23

# Environment & Config
python-dotenv==1.0.0

# Async Support
python-asyncio==3.4.3

# API Documentation & Validation
marshmallow==3.20.1

# Logging & Monitoring
python-json-logger==2.0.7

# Error Handling
sentry-sdk==1.38.0

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1

# Production Server
gunicorn==21.2.0
waitress==2.1.2

# Flask
Flask==2.3.0
Flask-CORS==4.0.0
PyJWT==2.8.0
oracledb==1.3.0